6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /**
   * main - Print Message
-  * return: 0 (success)
+  * Return: 0
   */
 int main(void)
 {
@@ -11,11 +11,10 @@ int main(void)
 	long long int d;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)
-	sizeof(a));
-	printf("Size of an int: %lu byte(s)\n", sizeof(b));
-	printf("Size of a long in: %lu byte(s)\n", sizeof(c);
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(d)
+	printf("Size of a char: %d byte(s)\n", sizeof(a));
+	printf("Size of an int: %d byte(s)\n", sizeof(b));
+	printf("Size of a long in: %d byte(s)\n", sizeof(c));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(d))
 	printf("Size of float: %d byte(s)\n", sizeof(f));
 	return (0);
 }
